Further patched jump table detection for 64bit. Now we try recursive function disassembly, falling back to linear disassembly when either a (bad) instruction or an undetectable jump table is found.
git-svn-id: svn+ssh://bono.sdsc.edu/projects/pmac/svnrepos/pmac/trunk/X86ElfInstrumentor@2553 55925d41-9515-0410-9586-c6f35bcc715f

+++ include/RawSection.h
@@ -15,18 +15,18 @@ class SectionHeader;
 class DataReference : public Base {
 private:
     uint64_t data;
-    uint32_t sectionOffset;
+    uint64_t sectionOffset;
     bool is64bit;
 
     RawSection* rawSection;
     AddressAnchor* addressAnchor;
 
 public:
-    DataReference(uint64_t dat, RawSection* rawsect, bool is64, uint32_t off);
+    DataReference(uint64_t dat, RawSection* rawsect, bool is64, uint64_t off);
     ~DataReference();
 
     uint64_t getBaseAddress(); 
-    uint32_t getSectionOffset() { return sectionOffset; }
+    uint64_t getSectionOffset() { return sectionOffset; }
     void initializeAnchor(Base* link);
     AddressAnchor* getAddressAnchor() { return addressAnchor; }
